Sequence Through Taxonomic Metrics

This function is a wrapper for the other "taxa_" functions that sequences through unique values within specified columns. The intent is to provide the ability to quickly calculate metrics for a large number of taxa without having to call a "taxa_" function for each taxon of interest.

taxa_seq(
  .dataframe,
  .key_col,
  .counts_col,
  .filter_cols_vec,
  .group_col,
  .unnest_col = NULL,
  .job,
  .base_log = 2,
  .q = NULL
)

Arguments

.dataframe
A data frame where each row should represent the number of individuals enumerated for a single taxon collected during a single sampling event.

.key_col
One unquoted column name that represents a key (i.e., unique ID) for a sampling event for which to group (i.e., aggregate) the data.

.counts_col
One unquoted column name that represents taxonomic counts.

.filter_cols_vec
A quoted vector of column names for which the function will sequence through each unique value to perform the specified .job.

.group_col
One unquoted column name that represents a taxomic rank or group of interest.

.unnest_col
One unqouted column name that represents nested data. If this column is NULL (default), then the data will not be unnested.

.job
A character string specifying the metric of interest. Below is a list of exceptable inputs:

"abund"Description needed

"pct"Description needed

"rich"Description needed

"pct_rich"Description needed

"shannon"Description needed

"effective_shannon"Description needed

"simpson""Description needed

"invsimpson"Description needed

"gini_simpson"Description needed

"effective_simpson"Description needed

"pielou"Description needed

"margalef"Description needed

"menhinick"Description needed

"hill"Description needed

"renyi"Description needed

.base_log
The base log value used during the calculation of Shannon Diversity index ("shannon") or Effective Shannon Diversity ("effective_shannon"). The default value is two.

.q
The exponent used during the calculation of Hill Numbers ("hill") and Renyi Entropy ("renyi").

Value

A data frame where each column represents numeric metric values.